+++ fabric/src/main/java/com/dudko/blazinghot/data/fabric/BlazingLangGen.java
@@ -1,5 +1,6 @@
 package com.dudko.blazinghot.data.fabric;
 
+import com.dudko.blazinghot.content.fluids.MoltenMetal;
 import com.dudko.blazinghot.registry.BlazingTags;
 import com.dudko.blazinghot.registry.CommonTags;
 import com.google.gson.JsonElement;
@@ -17,6 +18,7 @@ public class BlazingLangGen {
 
         BlazingTags.provideLangEntries(langConsumer);
         CommonTags.provideLangEntries(langConsumer);
+        MoltenMetal.provideLangEntries(langConsumer);
         provideDefaultLang("tooltips", langConsumer);
         provideDefaultLang("interface", langConsumer);
     }