+++ fabric/src/main/java/com/dudko/blazinghot/data/recipe/fabric/BlazingCompactingRecipeGen.java
@@ -1,17 +1,12 @@
 package com.dudko.blazinghot.data.recipe.fabric;
 
-import com.dudko.blazinghot.multiloader.MultiFluids;
 import com.simibubi.create.AllRecipeTypes;
 import com.simibubi.create.foundation.recipe.IRecipeTypeInfo;
 import net.minecraft.data.PackOutput;
-import net.minecraft.tags.TagKey;
-import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.material.Fluid;
 
 import java.util.List;
 
 import static com.dudko.blazinghot.content.fluids.MoltenMetal.ALL_METALS;
-import static com.dudko.blazinghot.data.recipe.fabric.Ingredients.*;
 
 @SuppressWarnings("unused")
 public class BlazingCompactingRecipeGen extends BlazingProcessingRecipeGen {
@@ -21,7 +16,11 @@ public class BlazingCompactingRecipeGen extends BlazingProcessingRecipeGen {
     }
 
     List<GeneratedRecipe> ALL_MOLTEN_COMPACTING_RECIPES = ALL_METALS.stream()
-            .map(metal -> moltenToIngot(metal.moltenName(), metal.fluidTag(), metal.ingot().get()))
+            .map(metal -> create(metal.moltenName(),
+                    b -> custom(b).blazinghot$convertMeltables()
+                            .blazinghot$finish()
+                            .require(metal.fluidTag(), metal.compactingResult().getSecond())
+                            .output(metal.compactingResult().getFirst())))
             .toList();
 
     @Override
@@ -29,11 +28,4 @@ public class BlazingCompactingRecipeGen extends BlazingProcessingRecipeGen {
         return AllRecipeTypes.COMPACTING;
     }
 
-    private GeneratedRecipe moltenToIngot(String name, TagKey<Fluid> tag, ItemLike result) {
-        return create(name,
-                b -> custom(b).blazinghot$convertMeltables()
-                        .blazinghot$finish()
-                        .require(tag, MultiFluids.Constants.INGOT.platformed())
-                        .output(result));
-    }
 }