+++ fabric/src/main/java/com/dudko/blazinghot/registry/fabric/BlazingFluidsImpl.java
@@ -22,7 +22,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.FluidState;
@@ -30,6 +29,7 @@ import net.minecraft.world.level.material.FluidState;
 import javax.annotation.Nullable;
 import java.util.Arrays;
 import java.util.Iterator;
+import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.function.Function;
 
@@ -60,7 +60,9 @@ public class BlazingFluidsImpl {
     private static FluidEntry<SimpleFlowableFluid.Flowing> createFromLava(String name, int tickRate, int decreaseRate) {
         return REGISTRATE
                 .standardFluid(name)
-                .tag(fluidTagOf(name, COMMON), fluidTagOf(name, FORGE)) // replace this with something else if the datagen fails to generate these tags
+                .tag(fluidTagOf(name, COMMON),
+                        fluidTagOf(name,
+                                FORGE)) // replace this with something else if the datagen fails to generate these tags
                 .tag(FluidTags.LAVA) // fabric: lava tag controls physics
                 .fluidProperties(p -> p
                         .levelDecreasePerBlock(decreaseRate)
@@ -116,10 +118,14 @@ public class BlazingFluidsImpl {
     @Nullable
     public static BlockState getLavaInteraction(FluidState fluidState, FluidState metFluidState) {
         Fluid fluid = fluidState.getType();
-        Fluid metFluid = metFluidState.getType();
 
-        if (fluid.isSame(MOLTEN_METALS.getFluid(MoltenMetal.BLAZE_GOLD)) && metFluidState.is(FluidTags.WATER)) {
-            return Blocks.NETHERRACK.defaultBlockState();
+        for (MoltenMetal metal : MoltenMetal.ALL_METALS) {
+            Map<Fluid, BlockState> interactions = metal.fluidInteractions;
+
+            for (Map.Entry<Fluid, BlockState> entry : interactions.entrySet()) {
+                if (fluid.isSame(MOLTEN_METALS.getFluid(metal)) && metFluidState.is(entry.getKey()))
+                    return entry.getValue();
+            }
         }
         return null;
     }