+++ changelog.md
@@ -1,14 +1,19 @@
 ------------------------------------------------------
-Create: Blazing Hot 0.5.1
+Create: Blazing Hot 0.6
 ------------------------------------------------------
 
 ### Added:
+* New fluids: Molten Zinc, Copper, Brass, Ancient Debris, Netherite, Andesite
 * Crafting recipe for Blaze Apple
 
 ### Changed:
 * Molten Blaze Gold source blocks now turn into Obsidian when touching water
 * Modern Lamp Panels can now be waterlogged
 * Increased chance of getting Blaze Powder when crushing Blaze Gold Rods from 30% to 50%
+* Several recipes now use item tags instead of specific items
+
+### Removed:
+* Metal Block melting recipes
 
 ### Fixed:
 * Fixed mod .jar files using `blazinghot` name instead of `create_blazing_hot`